requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.6
invoke==2.2.0
paramiko==4.0.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.1
